pyproject.toml
[project]
name = "final-assignmetn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "streamlit>=1.54.0",
    "tensorflow>=2.20.0",
]
